\name{twitterFUN-package}
\alias{twitterFUN-package}
\alias{twitterFUN}
\docType{package}
\title{
\packageTitle{twitterFUN}
}
\description{
\packageDescription{twitterFUN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{twitterFUN}
\packageIndices{twitterFUN}
}
\author{
\packageAuthor{twitterFUN}

Maintainer: \packageMaintainer{twitterFUN}
}
\keyword{ package }
